pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/phlo"]
include = ["src/phlo/plugins/registry_data.json"]

[tool.hatch.build.targets.sdist]
include = ["src/phlo/plugins/registry_data.json", "registry/schema/v1.json", "registry/plugins.json"]

[project]
name = "phlo"
version = "0.1.0-alpha.1"
description = "Lakehouse platform"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dagster",
    "dagster-webserver",
    "pydantic-settings",
    "pandas",
    "pandera",
    "requests",
    "psycopg2-binary",
    "asyncpg>=0.30.0",
    "click>=8.0",
    "rich>=13.0",
    "duckdb>=0.9.0",
    "python-ulid>=2.0.0",
    "httpx>=0.28.1",
    "fastapi>=0.127.0",
]

[project.optional-dependencies]
core-services = [
    "phlo-dagster>=0.1.0",
    "phlo-postgres>=0.1.0",
    "phlo-minio>=0.1.0",
    "phlo-nessie>=0.1.0",
    "phlo-trino>=0.1.0",
]
# Convenience alias - install all default services with: pip install phlo[defaults]
defaults = [
    "phlo-core-plugins>=0.1.0",
    "phlo-iceberg>=0.1.0",
    "phlo-dlt>=0.1.0",
    "phlo-dbt>=0.1.0",
    "phlo-dagster>=0.1.0",
    "phlo-postgres>=0.1.0",
    "phlo-minio>=0.1.0",
    "phlo-nessie>=0.1.0",
    "phlo-trino>=0.1.0",
]
openmetadata = [
    "phlo-openmetadata>=0.1.0",
]

[project.scripts]
phlo = "phlo.cli.main:cli"

# Plugin entry point groups
# External plugins can register themselves in these groups
[project.entry-points."phlo.plugins.sources"]
# Example: my_source = "my_package.my_module:MySourceConnector"

[project.entry-points."phlo.plugins.quality"]
# Example: my_check = "my_package.my_module:MyQualityCheck"

[project.entry-points."phlo.plugins.transforms"]
# Example: my_transform = "my_package.my_module:MyTransform"

[tool.uv.workspace]
members = [".", "packages/*"]

[tool.uv.sources]
phlo = { workspace = true }
phlo-core-plugins = { workspace = true }
phlo-iceberg = { workspace = true }
phlo-dlt = { workspace = true }
phlo-dbt = { workspace = true }
phlo-dagster = { workspace = true }
phlo-api = { workspace = true }
phlo-quality = { workspace = true }
phlo-lineage = { workspace = true }
phlo-alerting = { workspace = true }
phlo-observatory = { workspace = true }
phlo-postgres = { workspace = true }
phlo-minio = { workspace = true }
phlo-nessie = { workspace = true }
phlo-trino = { workspace = true }
phlo-openmetadata = { workspace = true }
phlo-metrics = { workspace = true }
phlo-testing = { workspace = true }

[dependency-groups]
dev = [
    "pytest>=8.4.2",
    "sqlfluff>=3.5.0",
    "phlo-dlt>=0.1.0",
    "phlo-iceberg>=0.1.0",
    "phlo-dbt>=0.1.0",
    "phlo-nessie>=0.1.0",
    "phlo-openmetadata>=0.1.0",
    "phlo-metrics>=0.1.0",
    "phlo-testing>=0.1.0",
    "ruff",
    "basedpyright",
    "ty",
]

[tool.pytest.ini_options]
testpaths = ["tests", "packages/*/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
norecursedirs = ["src", ".venv"]
# Skip integration tests by default - they require dbt manifest and running services
# Run integration tests with: pytest -m integration
addopts = "--ignore=src --continue-on-collection-errors -m 'not integration'"
markers = [
    "integration: marks tests as requiring external services or dbt manifest (run with '-m integration')",
]
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]  # Line too long - let the formatter handle this

[tool.ruff.lint.per-file-ignores]
# CLI main.py imports commands after group definition (intentional pattern)
"src/phlo/cli/main.py" = ["E402"]
# Fixtures imports at bottom for conftest template
"src/phlo/testing/fixtures.py" = ["E402"]
# Test files have pytestmark before imports (required for pytest collection)
"tests/test_definitions.py" = ["E402"]
"tests/test_system_e2e.py" = ["E402"]
"packages/phlo-dbt/tests/test_transform.py" = ["E402"]
"packages/phlo-iceberg/tests/test_resources.py" = ["E402"]
"tests/test_quality.py" = ["E402"]
"tests/test_rest_api.py" = ["E402"]
"tests/test_graphql_api.py" = ["E402"]
"packages/phlo-iceberg/tests/scripts/test_duckdb_iceberg.py" = ["E402"]

[tool.basedpyright]
include = ["src/phlo"]
exclude = ["src/phlo/testing"]  # Mock/placeholder code with dynamic imports
stubPath = "typings"
pythonVersion = "3.11"
typeCheckingMode = "standard"  # Less strict than "strict" mode
# Disable warnings that are too noisy with pandas/dagster/dynamic imports
reportUnknownMemberType = false
reportUnknownArgumentType = false
reportUnknownVariableType = false
reportUnknownParameterType = false
reportMissingTypeStubs = false
reportPrivateUsage = false
reportPossiblyUnboundVariable = false
reportRedeclaration = false
# Common issues in codebases using dynamic typing (pandas, dagster, etc.)
reportOptionalMemberAccess = false
reportOptionalSubscript = false
reportArgumentType = false  # Many false positives with pandas/dict operations - TODO: Re-enable incrementally
reportCallIssue = false  # TODO: Re-enable incrementally
reportAttributeAccessIssue = false  # TODO: Re-enable incrementally
reportGeneralTypeIssues = false  # TODO: Re-enable incrementally
reportIncompatibleVariableOverride = false  # TODO: Re-enable incrementally
